Refactored ability interaction
