+++ ProjectWeather/Assets/Prefabs/Ice.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c4a23f3720d1fc14696a2d504b365153
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
