--- ProjectWeather/Assets/Scripts/Abilities/Freezeable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 95f6169933cbf8846bbd0fa6d81278ec
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
