--- ProjectWeather/Assets/Scripts/Abilities/IceGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e097aba8622cdbf47a02ff01f6b4fa2b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
