--- ProjectWeather/Assets/Scripts/Abilities/Rainable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 514196092d4cb184d97531ab6137dc52
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
