--- ProjectWeather/Assets/Scripts/Abilities/Sunable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6fbf49af5b09a6c4580c31603aa8c4c6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
