--- ProjectWeather/Assets/Scripts/Abilities/WaterOnFreeze.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 42fc7b37eb688a64e8954f6ee60a71c3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
