--- ProjectWeather/Assets/Scripts/Abilities/WaterOnSun.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 68ffa9b382e11104e84fdc321681e124
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
