--- ProjectWeather/Assets/Scripts/Abilities/Windable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b95814b1ffaff0f4490e1f2048e641f1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
