+++ ProjectWeather/Assets/Scripts/sunKey/sunKeyOnSun3.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class sunKeyOnSun3 : Sunable
+public class sunKeyOnSun3 : Interactible
 {
     public Light keyLight;
     // Start is called before the first frame update
